Rename membership package to `developmentmembership` to give context on when its used
