+++ internal/developmentmembership/membership_test.go
@@ -1,4 +1,4 @@
-package membership
+package developmentmembership
 
 import (
 	"sort"
